Adjusted code to be compatible with new parameter data file tables. Lookup now uses column 6 as lower case base change check, 	CC.1 to CC.64 is now just -CC in lookup table (still -CC.[0-9]* in bam/fastq files. 	degenerate bases are upcase reference, thus only base changes will be tested for in NT
To Do: CC are not part of recominant poplulation, but NT with CC or IS
with CC basechanges will be removed to recombo pool.  This means that
recombinants which can be NT/IS, NT/CC, IS/CC, the latter two
populations are not correctly counted and the current NT/IS logic to
adjust viral load and runQC is not optimal.